Use binary storage format in o0keyChainStore
From Qt documentation: "The QDataStream class provides serialization of
binary data to a QIODevice."
When this binary data was converted to UTF-8, it often started with
a few null (zero) bytes. The qtkeychain library then stored it as
an empty string (tested on GNU/Linux).

Storing the data in the binary format makes o0keyChainStore work.