Fix authorization failure with 127.0.0.1
This failure can be reproduced only on Windows machine with
Microsoft Edge as a default browser. For some unknown reason
the browser can't connect to the local server if the IP address
is 127.0.0.1, but it connects properly to localhost.

I've verified that localhost works well for SmugMug Authorization
on Windows, Mac, Linux with Qt 5.4.0.

In commit 26c52593f81e7470e594a6c8e8b026a6db609c13 the 127.0.0.1
address was changed to localhost because of a Mac OS X issue. But in the next
commit (bea2c601f2cf4007c0ad3d756b70be8cd33b88a1) the 127.0.0.1
address was made the default again without an explanation.