Add possibility to use an external web interceptor to receive the OAuth tokens instead of creating an O2ReplyServer
In order to catch the OAuth callback, O2 creates a local web server running with the registered OAuth callback port. While this generally works, there are some cases where this causes problems:
- there is already on the machine a local server running with the specified port
- there might be some firewall, antivirus and other tools that block the local web server.

To solve these issues, it is now possible to use an external web interceptor in O2 instead of creating an O2ReplyServer.

Each O0BaseAuth subclasses (O1, O2, O2Facebook, ...) now have a new parameter 'inUseExternalInterceptor' in their constructor. By default this is set to false and use the old O2ReplyServer mechanism.

If you set this parameter to true, the O2ReplyServer won't be created. Instead the application using the O2 library has to intercept itself the OAuth callback and then call the new O2 function processOAuthCallbackFromExternalInterceptor(const QString &inURLString). The O2 library will then finalize the OAuth setup.

------------------------------------------------------------------------

Here is an example on how you can reproduce such an issue on macOS:

1. Open a Terminal window and type python -m SimpleHTTPServer 8888.
This creates a local web server running on the port 8888. You can open in your favorite webbrowser the page http://127.0.0.1:8888 and confirm that you see the content of your home.
2. Compile and run the example 'facebookdemo' and try to link a Facebook account

Result: The local web server will catch the OAuth callback URL and adding the OAuth account will fail in 'facebookdemo'.

------------------------------------------------------------------------

The O2 changes also include a new example project called 'facebookexternalinterceptordemo'. It does the same as the 'facebookdemo' example except that it doesn't create a local web server with O2ReplyServer. Instead when 'FBDemo::onOpenBrowser(const QUrl &url)' is called, it creates a web window with a web interceptor mechanism. When you enter the Facebook credentials, the callback url is catched in the web window and the code is passed to the O2 library to finalize the OAuth linking.

1. Open a Terminal window and type python -m SimpleHTTPServer 8888.
This creates a local web server running on the port 8888. You can open in your favorite webbrowser the page http://127.0.0.1:8888 and confirm that you see the content of your home.
2. Compile and run the example 'facebookexternalinterceptordemo' and try to link a Facebook account

Result: Note that you are able to link to your Facebook account.