Some OAuth 2 services require the access token to be sent as a Authentication HTTP header. This is the case for Twitch and Mixer which require respectively an 'OAuth' and 'Bearer' header for OAuth 2.
For those special cases, O2Requestor fails when retrying because the Authentication HTTP header is not updated with the new access token: when the access token expires and is refreshed, O2Requestor::retry() needs to update the Authentication HTTP header. In order to do so, O2Requestor needs to know the format of the Authentication HTTP header.

This change adds an optional property 'accessTokenInAuthenticationHTTPHeaderFormat_' to O2Requestor to make it work which those services.